Add new login-form and registration-form components. Refactor login and registration pages
