+++ src/webclient/main/services/validator/index.js
@@ -20,10 +20,10 @@ export default class Validator {
         resolve();
       } else {
         if (login.length < minLength) {
-          reject(new VerificationError(422, 'login',
+          reject(new VerificationError('login',
             `Login should be longer than ${minLength} symbols.`));
         } else {
-          reject(new VerificationError(422, 'login',
+          reject(new VerificationError('login',
             'Login should contains only latin letters or numbers.'));
         }
       }
@@ -48,10 +48,10 @@ export default class Validator {
         resolve();
       } else {
         if (password.length < minLength) {
-          reject(new VerificationError(422, 'password',
+          reject(new VerificationError('password',
             `Password should be longer than ${minLength} symbols.`));
         } else {
-          reject(new VerificationError(422, 'password',
+          reject(new VerificationError('password',
             'Password should contain at least 1 uppercase and lowercase letters and 1 digit.'));
         }
       }
@@ -62,8 +62,8 @@ export default class Validator {
    *  Validates confirm password row if its fully fits the rules.
    *  Rules:
    *      1)passwords should be strict equal.
-   * @param {string} confirmPassword - form row that contains user confirmation password.
-   * @param {string} password - form row that contains user password.
+   * @param {string} confirmPassword - login-form row that contains user confirmation password.
+   * @param {string} password - login-form row that contains user password.
    * @return {Promise} returns true if confirmation password and password are strict equal,
    * if they are not:
    *      1)Renders warning message under the confirmation password input.
@@ -75,7 +75,7 @@ export default class Validator {
       if (confirmPassword === password) {
         resolve();
       } else {
-        reject(new VerificationError(422, 'confirmPassword', 'Passwords do not match.'));
+        reject(new VerificationError('confirmPassword', 'Passwords do not match.'));
       }
     }));
   }
